trial/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.3
python-socketio==5.13.0
python-engineio==4.12.2
motor==3.7.1
pymongo==4.13.2
openai==1.92.2
python-dotenv==1.1.1
requests==2.32.4
cachetools==6.1.0
bidict==0.23.1
thefuzz==0.22.1
python-dateutil
datetime
timezone
pydantic<2.0